personal_assistant/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
groq==0.4.1
python-dotenv==1.0.0
httpx>=0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
pytz>=2023.3
python-dateutil==2.8.2 